pyproject.toml
[tool.poetry]
name = "subtitlecore"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
webvtt-py = "^0.4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
parse2sens = "subtitlecore.parse2sens:run" 


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

